+++ example/calculator.dart
@@ -57,7 +57,7 @@ void _digit1(State<String> state) {
   final $pos = state.pos;
   while (state.pos < source.length) {
     final c = source.codeUnitAt(state.pos);
-    final ok = c >= 48 && c <= 57;
+    final ok = c <= 57 && c >= 48;
     if (!ok) {
       break;
     }
@@ -120,7 +120,7 @@ void _ws(State<String> state) {
   final source = state.source;
   while (state.pos < source.length) {
     final c = source.codeUnitAt(state.pos);
-    final ok = c <= 32 && (c >= 9 && c <= 10 || c == 13 || c == 32);
+    final ok = c == 13 || c <= 10 ? c >= 9 : c == 32;
     if (!ok) {
       break;
     }