+++ example/example.dart
@@ -15,7 +15,7 @@ void _ws(State<String> state) {
   final source = state.source;
   while (state.pos < source.length) {
     final c = source.codeUnitAt(state.pos);
-    final ok = c <= 32 && (c >= 9 && c <= 10 || c == 13 || c == 32);
+    final ok = c == 13 || c <= 10 ? c >= 9 : c == 32;
     if (!ok) {
       break;
     }
@@ -70,8 +70,11 @@ int? _escapeHex(State<String> state) {
     var $count = 0;
     while ($count < 4 && state.pos < source.length) {
       final c = source.codeUnitAt(state.pos);
-      final ok = c <= 102 &&
-          (c >= 48 && c <= 57 || c >= 65 && c <= 70 || c >= 97 && c <= 102);
+      final ok = c <= 57
+          ? c >= 48
+          : c <= 70
+              ? c >= 65
+              : c <= 102 && c >= 97;
       if (!ok) {
         break;
       }