+++ lib/src/sequence/sequence.dart
@@ -1,5 +1,293 @@
 part of '../../sequence.dart';
 
+class Sequence10<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  final ParserBuilder<I, dynamic> parser4;
+
+  final ParserBuilder<I, dynamic> parser5;
+
+  final ParserBuilder<I, dynamic> parser6;
+
+  final ParserBuilder<I, dynamic> parser7;
+
+  final ParserBuilder<I, dynamic> parser8;
+
+  final ParserBuilder<I, dynamic> parser9;
+
+  final ParserBuilder<I, dynamic> parser10;
+
+  const Sequence10(
+      this.parser1,
+      this.parser2,
+      this.parser3,
+      this.parser4,
+      this.parser5,
+      this.parser6,
+      this.parser7,
+      this.parser8,
+      this.parser9,
+      this.parser10);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [
+      parser1,
+      parser2,
+      parser3,
+      parser4,
+      parser5,
+      parser6,
+      parser7,
+      parser8,
+      parser9,
+      parser10
+    ];
+  }
+}
+
+class Sequence11<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  final ParserBuilder<I, dynamic> parser4;
+
+  final ParserBuilder<I, dynamic> parser5;
+
+  final ParserBuilder<I, dynamic> parser6;
+
+  final ParserBuilder<I, dynamic> parser7;
+
+  final ParserBuilder<I, dynamic> parser8;
+
+  final ParserBuilder<I, dynamic> parser9;
+
+  final ParserBuilder<I, dynamic> parser10;
+
+  final ParserBuilder<I, dynamic> parser11;
+
+  const Sequence11(
+      this.parser1,
+      this.parser2,
+      this.parser3,
+      this.parser4,
+      this.parser5,
+      this.parser6,
+      this.parser7,
+      this.parser8,
+      this.parser9,
+      this.parser10,
+      this.parser11);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [
+      parser1,
+      parser2,
+      parser3,
+      parser4,
+      parser5,
+      parser6,
+      parser7,
+      parser8,
+      parser9,
+      parser10,
+      parser11
+    ];
+  }
+}
+
+class Sequence2<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  const Sequence2(this.parser1, this.parser2);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [parser1, parser2];
+  }
+}
+
+class Sequence3<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  const Sequence3(this.parser1, this.parser2, this.parser3);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [parser1, parser2, parser3];
+  }
+}
+
+class Sequence4<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  final ParserBuilder<I, dynamic> parser4;
+
+  const Sequence4(this.parser1, this.parser2, this.parser3, this.parser4);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [parser1, parser2, parser3, parser4];
+  }
+}
+
+class Sequence5<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  final ParserBuilder<I, dynamic> parser4;
+
+  final ParserBuilder<I, dynamic> parser5;
+
+  const Sequence5(
+      this.parser1, this.parser2, this.parser3, this.parser4, this.parser5);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [parser1, parser2, parser3, parser4, parser5];
+  }
+}
+
+class Sequence6<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  final ParserBuilder<I, dynamic> parser4;
+
+  final ParserBuilder<I, dynamic> parser5;
+
+  final ParserBuilder<I, dynamic> parser6;
+
+  const Sequence6(this.parser1, this.parser2, this.parser3, this.parser4,
+      this.parser5, this.parser6);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [parser1, parser2, parser3, parser4, parser5, parser6];
+  }
+}
+
+class Sequence7<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  final ParserBuilder<I, dynamic> parser4;
+
+  final ParserBuilder<I, dynamic> parser5;
+
+  final ParserBuilder<I, dynamic> parser6;
+
+  final ParserBuilder<I, dynamic> parser7;
+
+  const Sequence7(this.parser1, this.parser2, this.parser3, this.parser4,
+      this.parser5, this.parser6, this.parser7);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [parser1, parser2, parser3, parser4, parser5, parser6, parser7];
+  }
+}
+
+class Sequence8<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  final ParserBuilder<I, dynamic> parser4;
+
+  final ParserBuilder<I, dynamic> parser5;
+
+  final ParserBuilder<I, dynamic> parser6;
+
+  final ParserBuilder<I, dynamic> parser7;
+
+  final ParserBuilder<I, dynamic> parser8;
+
+  const Sequence8(this.parser1, this.parser2, this.parser3, this.parser4,
+      this.parser5, this.parser6, this.parser7, this.parser8);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [
+      parser1,
+      parser2,
+      parser3,
+      parser4,
+      parser5,
+      parser6,
+      parser7,
+      parser8
+    ];
+  }
+}
+
+class Sequence9<I> extends _SequenceN<I> {
+  final ParserBuilder<I, dynamic> parser1;
+
+  final ParserBuilder<I, dynamic> parser2;
+
+  final ParserBuilder<I, dynamic> parser3;
+
+  final ParserBuilder<I, dynamic> parser4;
+
+  final ParserBuilder<I, dynamic> parser5;
+
+  final ParserBuilder<I, dynamic> parser6;
+
+  final ParserBuilder<I, dynamic> parser7;
+
+  final ParserBuilder<I, dynamic> parser8;
+
+  final ParserBuilder<I, dynamic> parser9;
+
+  const Sequence9(this.parser1, this.parser2, this.parser3, this.parser4,
+      this.parser5, this.parser6, this.parser7, this.parser8, this.parser9);
+
+  @override
+  List<ParserBuilder<I, dynamic>> _getParsers() {
+    return [
+      parser1,
+      parser2,
+      parser3,
+      parser4,
+      parser5,
+      parser6,
+      parser7,
+      parser8,
+      parser9
+    ];
+  }
+}
+
 abstract class _Sequence<I, O> extends ParserBuilder<I, O> {
   static const _template = '''
 final {{pos}} = state.pos;
@@ -31,15 +319,17 @@ if (state.ok) {
     }
 
     final fast = result == null;
-    final ignoreVoid = !_useResultsOfFastParsers();
     final values = context.allocateLocals(['pos']);
     final results = <ParserResult>[];
     var templateParsers = '{{next}}';
     for (var i = 0; i < parsers.length; i++) {
       final parser = parsers[i];
-      final type = parser.getResultType();
-      final isVoid = ignoreVoid ? type == 'void' : fast;
-      final r1 = context.getResult(parser, !isVoid);
+      var useResult = !fast;
+      if (!fast) {
+        useResult = _useParserResult(parser, i);
+      }
+
+      final r1 = context.getResult(parser, useResult);
       if (r1 != null) {
         results.add(r1);
       }
@@ -50,7 +340,7 @@ if (state.ok) {
         'var1': '{{var$index}}',
       };
       final templateParser =
-          render2(isVoid, _templateParserFast, _templateParser, values);
+          render2(!useResult, _templateParserFast, _templateParser, values);
       values.clear();
       values.addAll({
         'next': templateParser,
@@ -79,7 +369,19 @@ if (state.ok) {
 
   String _setResult(Context context, List<ParserResult> results);
 
-  bool _useResultsOfFastParsers() {
-    return true;
+  bool _useParserResult(ParserBuilder parser, int index);
+}
+
+abstract class _SequenceN<I> extends _Sequence<I, void> {
+  const _SequenceN();
+
+  @override
+  String _setResult(Context context, List<ParserResult> results) {
+    return ' //';
+  }
+
+  @override
+  bool _useParserResult(ParserBuilder parser, int index) {
+    return false;
   }
 }